fix: put route_online_debug into a separate file
